requirements.txt

python-telegram-bot==20.3
playwright==1.43.0
